Update build steps to work with arbitrary commits and updated submodules
Checking out a new commit or switching branches that update submodules
requires additional steps to checkout the submodules to that commit.

`git submodule update --init --recursive` takes care of that
(while the current README only worked for fresh clones).

The second change is that `mkdir -p` will succeed even if the directory
is already existent, thus making the whole recipe easier to just use as
copy-paste.